requirements.txt
python-binance
requests
pytest
Flask==2.2.2
flask-cors==3.1.1
requests==2.28.1

